% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sequence.R
\name{count_sequence}
\alias{count_sequence}
\title{Count occurrences of a sequence within a window}
\usage{
count_sequence(window, sequence)
}
\arguments{
\item{window}{A character string representing the window of DNA or RNA sequence to search within.}

\item{sequence}{A character string representing the sequence to count within the window.}
}
\value{
An integer value representing the total count of occurrences of the sequence within the window.
}
\description{
This function counts the number of occurrences of a given sequence within a specified window.
It uses the \code{countPattern} function to count the exact occurrences of the sequence in the window.
}
\examples{
# Example usage
window <- "AGCTAGCTAGCT"
sequence <- "AGC"
count_sequence(window, sequence)

}
\seealso{
\code{\link[Biostrings]{countPattern}} for more details on pattern matching.
}
