% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_windows.R
\name{create_windows}
\alias{create_windows}
\title{Create windows of specified size from the fasta file of a genome sequence}
\usage{
create_windows(genome, window_size = 1e+06)
}
\arguments{
\item{genome}{A character string representing the genome sequence to split into windows. Should be a fasta read in with readDNAStringSet}

\item{window_size}{A numeric value representing the size of each window. Default is 1e6 (1 million).}
}
\value{
A character vector where each element is a window (substring) of the genome sequence.
}
\description{
This function splits a genome sequence into windows of a specified size. It uses the \code{substring}
function to extract substrings from the genome at regular intervals defined by the \code{window_size} parameter.
}
\examples{
# Example usage
genome <- "ATGCATGCATGCATGCATGC"
create_windows(genome, window_size = 4)


}
