% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_table.R
\name{genome_table}
\alias{genome_table}
\title{Create a table of genomic features with calculations}
\usage{
genome_table(fasta_lengths_table, telomere_table, name, genome_size)
}
\arguments{
\item{fasta_lengths_table}{A data frame containing the FASTA length information. Columns must be "Chromosome" and "Length".}

\item{telomere_table}{A data frame containing telomere-related data, including "Chromosome" and counts of telomere repeats (e.g., "CCCTAAA_Counts" and "TTTAGGG_Counts"). Generated from function telomere_repeat_number}

\item{name}{A character string representing the name or identifier of the genome source.}

\item{genome_size}{A numeric value representing the total size of the genome.}
}
\value{
A data frame with additional columns related to telomere sizes and their normalization. The new columns include:
\describe{
\item{begin_telo_bp}{The length of the telomere at the beginning of the chromosome, in base pairs.}
\item{end_telo_bp}{The length of the telomere at the end of the chromosome, in base pairs.}
\item{begin_telo_start}{The start position of the beginning telomere, which is always 0.}
\item{begin_telo_end}{The end position of the beginning telomere, calculated based on the \code{begin_telo_bp}.}
\item{end_telo_start}{The start position of the ending telomere, calculated based on the chromosome length and \code{end_telo_bp}.}
\item{end_telo_end}{The end position of the ending telomere, which is equal to the chromosome length.}
\item{total_telo_bp}{The combined length of both telomeres in base pairs.}
\item{normalized_total_telo_size}{The total telomere size normalized by the genome size.}
\item{Source}{The name or identifier of the genome source (as provided by \code{name}).}
}
}
\description{
This function generates a table combining information from a FASTA length table and a telomere table,
while calculating additional telomere-related features (such as the size of the telomeres at both ends).
It also normalizes the total telomere size relative to the total genome size.
}
\examples{
# Example usage
fasta_lengths_table <- data.frame(Chromosome = c("Chr1", "Chr2"), Length = c(1000000, 2000000))
telomere_table <- data.frame(Chromosome = c("Chr1", "Chr2"), CCCTAAA_Counts = c(200, 300), TTTAGGG_Counts = c(150, 250))
genome_size <- 3000000
name <- "MyGenome"
genome_table(fasta_lengths_table, telomere_table, name, genome_size)

}
