% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggideoprimary.R
\name{ggideoprimary}
\alias{ggideoprimary}
\title{Generate an ideogram plot from a fasta file}
\usage{
ggideoprimary(
  path_fasta,
  chr_names = "Chr",
  tel_start_seq = "CCCTAAA",
  tel_end_seq = "TTTAGGG",
  size_windows = 1e+06,
  min_tel_count = 25,
  sample_name = NULL,
  title_plot = NULL,
  color_chr = "dodgerblue2",
  size_chr = 8,
  color_tel = "black",
  shape_tel = 16,
  scale_y = 1e-06,
  suffix_y_scale = "Mb",
  pos_legend = "bottom",
  size_legend = 0.25
)
}
\arguments{
\item{path_fasta}{A character string specifying the file path to the fasta file containing the genome sequence.}

\item{chr_names}{A character string specifying the pattern to identify chromosome names in the data. Default is "Chr".}

\item{tel_start_seq}{A character string representing the telomeric repeat sequence at the start of the chromosome. Default is "CCCTAAA".}

\item{tel_end_seq}{A character string representing the telomeric repeat sequence at the end of the chromosome. Default is "TTTAGGG".}

\item{size_windows}{A numeric value specifying the window size (in base pairs) for counting telomeric repeats. Default is \code{1e6} (1 Mb).}

\item{min_tel_count}{A numeric value representing the minimum number of telomeric repeats required to retain telomeric counts in the plot. Default is 25.}

\item{sample_name}{A character string specifying the name of the sample for labeling purposes. Default is \code{NULL}.}

\item{title_plot}{A character string specifying the title of the plot. Default is \code{NULL}.}

\item{color_chr}{A character string specifying the color of the chromosome segments in the plot. Default is "dodgerblue2".}

\item{size_chr}{A numeric value specifying the size (linewidth) of the chromosome segments. Default is 8.}

\item{color_tel}{A character string specifying the color of the telomere points. Default is "black".}

\item{shape_tel}{A numeric value specifying the shape of the telomere points. Default is 16 (filled circle).}

\item{scale_y}{A numeric value for scaling the y-axis. Default is \code{1e-6} to convert lengths to megabases (Mb).}

\item{suffix_y_scale}{A character string to append to the y-axis labels, usually a unit like "Mb". Default is "Mb".}

\item{pos_legend}{A character string specifying the position of the legend. Default is "bottom".}

\item{size_legend}{A numeric value specifying the size of the legend keys. Default is 0.25.}
}
\value{
A list containing:
\describe{
\item{genomic.table}{A data frame with chromosome, length, and telomere information.}
\item{ideogram}{A \code{ggplot} object representing the ideogram plot.}
}
}
\description{
This function creates an ideogram plot based on a genome sequence provided in a FASTA file. It reads the genome,
filters for specified chromosomes, calculates telomeric repeat counts, and visualizes the chromosome lengths with telomere regions marked.
The output includes a data table and a \code{ggplot} generated ideogram plot.
}
\examples{
# Example usage
# ggideoprimary(path_fasta = "path/to/genome.fasta", sample_name = "Sample_1", title_plot = "Genome Ideogram")

}
