% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_ideogram.R
\name{primary_ideogram}
\alias{primary_ideogram}
\title{Create an ideogram plot for primary genomic assemblies}
\usage{
primary_ideogram(
  genome.table,
  plot_title = NULL,
  x_axis_title = NULL,
  y_axis_title = "Chromosome Length",
  legend_title = "Telomere Length",
  chr_color = "dodgerblue2",
  chr_size = 8,
  tel_color = "black",
  tel_shape = 16,
  y_scale = 1e-06,
  y_scale_suffix = "Mb",
  legend_pos = "bottom",
  legend_size = 0.25
)
}
\arguments{
\item{genome.table}{A data frame containing genomic data, with columns for chromosome names, chromosome lengths,
and the start and end positions of telomere regions (\code{begin_telo_start}, \code{begin_telo_end}, \code{end_telo_start},
\code{end_telo_end}, \code{begin_telo_bp}, and \code{end_telo_bp}).}

\item{plot_title}{A character string specifying the plot title. Default is \code{NULL}, in which case no title is added.}

\item{x_axis_title}{A character string specifying the x axis title. Default is \code{NULL}, in which case no title is added.}

\item{y_axis_title}{A character string specifying the y axis title. Default is "Chromosome Length".}

\item{legend_title}{A character string specifying the legend title. Default is "Telomere Presence".}

\item{chr_color}{A character string specifying the color of the chromosome segments. Default is "dodgerblue2".}

\item{chr_size}{A numeric value specifying the size (linewidth) of the chromosome segments. Default is 8.}

\item{tel_color}{A character string specifying the color of the telomere points. Default is "black".}

\item{tel_shape}{A numeric value specifying the shape of the telomere points. Default is 16 (filled circle).}

\item{y_scale}{A numeric value for scaling the y-axis. Default is \code{1e-6} (for scaling the length to Mb).}

\item{y_scale_suffix}{A character string to append to the y-axis labels, usually a unit suffix like "Mb". Default is "Mb".}

\item{legend_pos}{A character string specifying the position of the legend. Default is "bottom".}

\item{legend_size}{A numeric value specifying the size of the legend keys. Default is 0.25.}
}
\value{
A \code{ggplot} object representing the ideogram plot with chromosome and telomere information visualized.
}
\description{
This function generates a basic ideogram plot of chromosomes, including the chromosome length and the telomere regions.
It visualizes the telomeres by marking the start and end points, and adjusting the plot appearance according to
specified parameters such as color, size, and scaling. This function was built to use datasets generated with genome_table
}
\examples{
# Example usage
genome.table <- data.frame(
  Chromosome = c("Chr1", "Chr2"),
  Length = c(100000000, 200000000),
  begin_telo_start = c(0, 0),
  begin_telo_end = c(5000000, 10000000),
  end_telo_start = c(95000000, 190000000),
  end_telo_end = c(100000000, 200000000),
  begin_telo_bp = c(5000000, 10000000),
  end_telo_bp = c(5000000, 10000000)
)
primary_ideogram(genome.table)

}
