% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telomere_repeat_number.R
\name{telomere_repeat_number}
\alias{telomere_repeat_number}
\title{Count the number of telomeric sequences repeated in a genome sequence}
\usage{
telomere_repeat_number(
  fasta,
  window = 1e+06,
  tel_start = "CCCTAAA",
  tel_end = "TTTAGGG"
)
}
\arguments{
\item{fasta}{A DNA sequences read in with the readDNAStringSet function.}

\item{window}{A numeric value specifying the size of the window (in base pairs) for splitting the genome sequences.
Default is 1e6 (1 million base pairs).}

\item{tel_start}{A character string representing the start sequence of the telomere. Default is "CCCTAAA".}

\item{tel_end}{A character string representing the end sequence of the telomere. Default is "TTTAGGG".}
}
\value{
A data table with the following columns:
\describe{
\item{Chromosome}{The name of the chromosome.}
\item{Window}{The window number (corresponding to the 1 Mb segments of the genome).}
\item{CCCTAAA_Counts}{The count of the start telomere repeat sequence ("CCCTAAA") in the window.}
\item{TTTAGGG_Counts}{The count of the end telomere repeat sequence ("TTTAGGG") in the window.}
}
}
\description{
This function counts the occurrences of specified telomere repeat sequences (start and end) within a genome sequence,
splitting the sequence into windows of a specified size (default is 1 Mb). It returns a data table with the counts
of telomere repeats for each window of each chromosome in the input genome.
}
\examples{
# Example usage with a list of genome sequences
fasta <- list(Chr1 = "ATGCATGCATGCATGCATGCATGC", Chr2 = "GGGCCCTAAAGGGCCCTAAA")
telomere_repeat_number(fasta, window = 4, tel_start = "CCCTAAA", tel_end = "TTTAGGG")

}
\seealso{
\code{\link{create_windows}} and \code{\link{count_sequence}} for the functions used to split sequences
and count the telomere repeats, respectively.
}
